[REFACTOR/#420] 패키지 정리 및 알림이 없을 경우 컴포넌트 처리
